uses more memory efficient hash value and exponent sizes

+++ src/basis.rs
@@ -97,12 +97,12 @@ impl Basis {
         // basis is the one of largest degree
         debug_assert!((self.previous_length..self.elements.len()).collect::<Vec<BasisLength>>()
                 .windows(2)
-                .all(|x| hash_table.monomials[self.elements[x[0]].monomials[0]].degree
-                    <= hash_table.monomials[self.elements[x[1]].monomials[0]].degree));
+                .all(|x| hash_table.monomials[self.elements[x[0]].monomials[0] as usize].degree
+                    <= hash_table.monomials[self.elements[x[1]].monomials[0] as usize].degree));
 
         // check for constant
         if hash_table.monomials[
-            self.elements[self.previous_length].monomials[0]].degree == 0 {
+            self.elements[self.previous_length].monomials[0] as usize].degree == 0 {
             self.is_constant = true;
         }
 
@@ -110,7 +110,7 @@ impl Basis {
         self.maximum_total_degree = max(
             self.maximum_total_degree,
             hash_table.monomials[self.elements[
-                self.elements.len()-1].monomials[0]].degree);
+                self.elements.len()-1].monomials[0] as usize].degree);
 
         // check redundancy due to resp. of new basis elements
         for i in self.previous_length..self.elements.len() {
